refactored api slice files and fixed imports
