+++ frontend/src/app/authApi.js
@@ -1,0 +1,23 @@
+import apiSlice from "./apiSlice";
+
+export const authApi = apiSlice.injectEndpoints({
+  endpoints: (builder) => ({
+    login: builder.mutation({
+      query: (body) => ({
+        url: "/auth/login",
+        method: "POST",
+        body,
+      }),
+    }),
+
+    register: builder.mutation({
+      query: (body) => ({
+        url: "/auth/register",
+        method: "POST",
+        body,
+      }),
+    }),
+  }),
+});
+
+export const { useLoginMutation, useRegisterMutation } = authApi;
