--- frontend/src/app/authApiSlice.js
@@ -1,35 +1,0 @@
-import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-
-export const authApiSlice = createApi({
-  reducerPath: "authApi",
-  baseQuery: fetchBaseQuery({
-    baseUrl: "http://localhost:3000/api/auth",
-    prepareHeaders: (headers, { getState }) => {
-      // get token
-      const token = getState().auth.token || localStorage.getItem("token");
-      if (token) {
-        headers.set("Authorization", `Bearer ${token}`);
-      }
-      return headers;
-    },
-  }),
-  endpoints: (builder) => ({
-    login: builder.mutation({
-      query: (body) => ({
-        url: "/login",
-        method: "POST",
-        body,
-      }),
-    }),
-
-    register: builder.mutation({
-      query: (body) => ({
-        url: "/register",
-        method: "POST",
-        body,
-      }),
-    }),
-  }),
-});
-
-export const { useLoginMutation, useRegisterMutation } = authApiSlice;
